feat: support load remote config in before method
