+++ packages/bootstrap/test/fixtures/mix-app/src/index.ts
@@ -1,0 +1,1 @@
+export * from './service/remoteConfigService';
