+++ packages/bootstrap/test/fixtures/mix-app/src/service/remoteConfigService.ts
@@ -1,0 +1,11 @@
+import { Provide, sleep } from '@midwayjs/decorator';
+
+@Provide()
+export class RemoteConfigService {
+  async getRemoteConfig() {
+    await sleep(1000);
+    return {
+      prepare: 'remote data',
+    }
+  }
+}
