+++ packages/core/src/context/requestContainer.ts
@@ -11,6 +11,7 @@ export class MidwayRequestContainer extends MidwayContainer {
     this.applicationContext = applicationContext;
     this.configService = this.applicationContext.getConfigService();
     this.environmentService = this.applicationContext.getEnvironmentService();
+    this.aspectService = this.applicationContext.getAspectService();
 
     // register ctx
     this.registerObject(REQUEST_CTX_KEY, ctx);
@@ -43,7 +44,7 @@ export class MidwayRequestContainer extends MidwayContainer {
     }
     if (this.registry.hasObject(identifier)) {
       const ins = this.registry.getObject(identifier);
-      return this.wrapperAspectToInstance(ins);
+      return this.aspectService.wrapperAspectToInstance(ins);
     }
     const definition = this.applicationContext.registry.getDefinition(
       identifier
@@ -58,7 +59,7 @@ export class MidwayRequestContainer extends MidwayContainer {
           definition,
           args,
         });
-        return this.wrapperAspectToInstance(ins);
+        return this.aspectService.wrapperAspectToInstance(ins);
       }
     }
 
@@ -76,7 +77,7 @@ export class MidwayRequestContainer extends MidwayContainer {
 
     if (this.registry.hasObject(identifier)) {
       const ins = this.registry.getObject(identifier);
-      return this.wrapperAspectToInstance(ins);
+      return this.aspectService.wrapperAspectToInstance(ins);
     }
 
     const definition = this.applicationContext.registry.getDefinition(
@@ -92,7 +93,7 @@ export class MidwayRequestContainer extends MidwayContainer {
           definition,
           args,
         });
-        return this.wrapperAspectToInstance(ins);
+        return this.aspectService.wrapperAspectToInstance(ins);
       }
     }
